Refactor Order, and VanquishDataGenerator
Most of this was removing old code for loading into HDFS and Hive.
Decoupling timestamp genration allowed for the creation of data trends
with it. I also ended up changing Order.scala to create orders with a
single product instead of many. A lot of this was just simplifiying our
code and making sure the changes can interacted with the old code. Has
not been test yet but that is for tomorrow.